+++ app/middlewares/index.js
@@ -1,0 +1,16 @@
+const { checkSchema } = require('express-validator');
+const { validationResult } = require('express-validator');
+const logger = require('../logger');
+const { schemaError } = require('../errors');
+
+const validateResult = (req, _, next) => {
+  const err = validationResult(req);
+  if (!err.isEmpty()) {
+    logger.error(`An error ocurred: ${JSON.stringify(err)}`);
+    return next(schemaError(err.mapped()));
+  }
+  return next();
+};
+const verifySchema = schema => checkSchema(schema);
+
+exports.validateBySchema = schema => [verifySchema(schema), validateResult];
