+++ app/routes.js
@@ -1,9 +1,10 @@
 const { healthCheck } = require('./controllers/healthCheck');
 const userController = require('./controllers/users');
 const { userSchema } = require('./schemas/users');
-const { validateBySchema, existUserDB } = require('./middlewares/users');
+const { existUserDB } = require('./middlewares/users');
+const { validateBySchema } = require('./middlewares');
 
 exports.init = app => {
   app.get('/health', healthCheck);
-  app.post('/users', [existUserDB, validateBySchema(userSchema)], userController.createUser);
+  app.post('/users', [validateBySchema(userSchema), existUserDB], userController.createUser);
 };
