+++ config/error.js
@@ -9,7 +9,7 @@ const ErrorMessages = {
   PASSWORD_REQUIRED: 'The password is required.',
   PASSWORD_NOT_LENGTH: 'Password should be at least 8 characters',
   USER_EMAIL_NOT_COMPANY: 'The email is not pertain to company.',
-  USER_EMAIL_EXITS: 'This email already exists',
+  USER_EMAIL_EXIST: 'This email already exist.',
   PASSWORD_NOT_COMPLY: 'The password not comply the required.',
   PASSWORD_NOT_ENCRYPT: 'The password encrypt have an error.'
 };
